Move FluentPathProvider into Connectors and drop IFluentSegmentSource
The IFluentSegmentSource seam existed only so a Registry-resident FluentPathProvider
could read a Connectors-resident FluentMappingRegistry. Nothing in Registry implements
or consumes it, so move FluentPathProvider into Connectors (next to the registry) as a
generic FluentPathProvider<TMetadata> that takes the registry directly, and remove the
interface. FluentMappingRegistry keeps TryGetSegment as a plain method.

The generic parameter only types the registry reference (the provider is metadata-agnostic)
and is confined to the two connector mapper constructors, so it never reaches consumers.
Registry returns to the general-purpose path engine; the MQTT and OPC UA public APIs are
unchanged. Update the Registry and Connectors API snapshots.
